Move test ui to mpp demo
